requirements.txt
langchain-community
pypdf
fastapi
uvicorn
pydantic
langchain
langchain_openai
langchain-community
langgraph
elevenlabs
python-dotenv
langchain_google_genai